Moved guiservthread_free_space_check method to GUIUtility (since it is in essence part of updating the GUI)
